+++ abstractions/dotnet/src/Method.cs
@@ -1,4 +1,4 @@
-// ------------------------------------------------------------------------------
+﻿// ------------------------------------------------------------------------------
 //  Copyright (c) Microsoft Corporation.  All Rights Reserved.  Licensed under the MIT License.  See License in the project root for license information.
 // ------------------------------------------------------------------------------
 
@@ -7,7 +7,7 @@ namespace Microsoft.Kiota.Abstractions
     /// <summary>
     ///     Represents the HTTP method used by a request.
     /// </summary>
-    public enum HttpMethod
+    public enum Method
     {
         /// <summary>
         ///     The HTTP GET method.